Rename src/ -> pythonturtle/, pythonturtle.py -> application.py
Use relative imports, obsolete chdir
