+++ pythonturtle/__init__.py
@@ -2,8 +2,6 @@
 An educational environment for learning Python, suitable for beginners
 and children. Inspired by LOGO.
 
-Call `run()` to run it.
-
 PythonTurtle aims to be the most low-threshold way for learning or
 teaching Python. The user is given command of an interactive Python
 shell, similar to IDLE, and is able to use Python functions to move a
@@ -11,8 +9,5 @@ turtle that is displayed on the screen. An illustrated help screen
 demonstrates how to move the turtle and introduces the student to the
 basics of Python programming.
 """
-
-from pythonturtle import run
-
-if __name__ == "__main__":
-    run()
+__version__ = '0.3.0'
+__url__ = 'http://pythonturtle.org'
