+++ pythonturtle/__main__.py
@@ -1,0 +1,7 @@
+"""
+Main entry point for executing the PythonTurtle application.
+"""
+from .application import run
+
+if __name__ == "__main__":
+    run()
