+++ pythonturtle/resources/about.txt
