+++ pythonturtle/resources/developers.txt
