+++ pythonturtle/resources/lets_code.png
