+++ pythonturtle/resources/turtle.png
