+++ pythonturtle/shelltoprocess/console.py
