+++ pythonturtle/turtlewidget.py
@@ -1,11 +1,11 @@
 """
-TurtleWidget is defined in this module, see its documentation.
+A TurtleWidget display the turtle and all the drawings that it made.
 """
 import wx
 
-from helpers import dump_queue, from_resource_folder
-from my_turtle import BITMAP_SIZE, Turtle, from_my_pos, from_my_angle
-from vector import Vector
+from .helpers import dump_queue, from_resource_folder
+from .my_turtle import BITMAP_SIZE, Turtle, from_my_pos, from_my_angle
+from .vector import Vector
 
 
 class TurtleWidget(wx.Panel):
